Fix checking if command is RAW
Originally, a command was determined to be raw if it was all uppercase and only alphanumerical. This breaks any command that contains a different character such as setting the temporary subwoofer level with SWL+05. A regular expression now checks if a command is all uppercase, contains `-` or `+`, and if it is alphanumerical. To achieve support for True/False, the entire string is checked using `^` and `$`. Thus, a command will be seen as raw if it starts with 1 or more uppercase letters, contains 0 or more `-` or `+` and ends with 0 or more numbers or uppercase letters in any combination.